# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food-Delivery",
    "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
    
      "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
  "/api/mobile/company/gallery/add": {
    "post": {
      "x-swagger-router-controller": "/api/mobile/company/gallery/add",
      "operationId": "index",
      "tags": ["Add  Gallery images"],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
        { "name": "companyId",  "in": "formData",  "required":"true","type":"string"},
        { "name": "mediaType",  "required":"true", "in": "formData","type":"string" ,"description":"send 'photo' to upload photos,'video' to uplaod videos"},
        { "name": "title",  "required":"true", "in": "formData", "type":"string"},
        { "name": "description", "required":"true", "in": "formData","type":"string"}


   ],

 
  
  
      "responses": {}
    }
  },
...
   "/api/mobile/cart/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/cart/add",
       "operationId": "index",
       "tags": ["Add to Cart"],
       "description": "Add Item to Cart",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
         { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "deliveryType",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "companyId",  "in": "formData",  "required":true,"type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
...
   "/api/mobile/address/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/address/list",
       "operationId": "index",
       "tags": ["Add New Address"],
       "description": "get Prduct listing by Category ID",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "addressName",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "latitude",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "longitude",  "in": "formData",  "required":true ,"type":"string" },
         { "name": "city",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "default",  "in": "formData","type":"integer"},
         { "name": "addressType",  "in": "formData"},
         { "name": "town",  "in": "formData"},
         { "name": "landmark",  "in": "formData"},
         { "name": "houseNo",  "in": "formData"}


    ],

  
   
   
       "responses": {}
     }
   },
...
   "/api/mobile/favourite/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/favourite/add",
       "operationId": "index",
       "tags": ["Add to Favourite"],
       "description": "get cart listing",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
         { "name": "serviceId",  "in": "formData",  "required":"true" ,"type":"string"},
         { "name": "companyId",  "in": "formData",  "required":true ,"type":"string"}



    ],

  
   
   
       "responses": {}
     }
   },
...
   "/api/mobile/tiffin/cart/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/add",
       "operationId": "index",
       "tags": ["Add Tiffin to cart "],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "tiffinId",  "in": "formData", "type":"string","required":true},
         { "name": "orderPrice",  "in": "formData", "type":"string","required":true},
         { "name": "orderTotalPrice",  "in": "formData", "type":"string","required":true},
         { "name": "companyId",  "in": "formData", "type":"string","required":true},
         { "name": "quantity",  "in": "formData", "type":"string","required":true},
         { "name": "fromDate",  "in": "formData", "type":"string","required":true},
         { "name": "package",  "in": "formData", "type":"string","required":true},
         { "name": "excPrice",  "in": "formData", "type":"string","description":"Excluded Total price"},
         { "name": "excAvaialbility",  "in": "formData", "type":"string","description":"Excluded Avaialbility if any like breakfast,lunch,dinner"},
         { "name": "excDays",  "in": "formData", "type":"string","description":"Excluded days like: monday,tuesday"}

        ],
       "responses": {}
     }
   },
...
   "/api/mobile/recipe/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/recipe/add",
       "operationId": "index",
       "tags": ["Add Recipes "],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "tempId",  "in": "formData", "type":"string","required":true,"description":"Send auto generated UUID"},
         { "name": "title",  "in": "formData", "type":"string","required":true},
         { "name": "description",  "in": "formData", "type":"string","required":true},
         { "name": "categoryId",  "in": "formData", "type":"string","required":true},
         { "name": "latitude",  "in": "formData", "type":"string"},
         { "name": "longitude",  "in": "formData", "type":"string"},
         { "name": "media",  "in": "formData","type":"array"  ,   "collectionFormat": "multi",
           "items": {
             "type": "string"
            }}         ],
       "responses": {}
     }
   },
   "/api/mobile/recipe/category/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/recipe/category/add",
       "operationId": "index",
       "tags": ["Add Category "],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "name",  "in": "formData", "type":"string","required":true}

        ],
       "responses": {}
     }
   },
...
   "/api/mobile/recipe/comment/add": {
    "post": {
      "x-swagger-router-controller": "/api/mobile/recipe/comment/add",
      "operationId": "index",
      "tags": ["Add Comment "],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
        { "name": "recipeId",  "in": "formData", "type":"string","required":true},
        { "name": "comment",  "in": "formData", "type":"string","required":true }

       ],
      "responses": {}
    }
  },
  "/api/mobile/recipe/comment/delete": {
    "delete": {
      "x-swagger-router-controller": "/api/mobile/recipe/comment/delete",
      "operationId": "index",
      "tags": ["Delete Comment "],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
        { "name": "commentId",  "in": "query", "type":"string","required":true}

       ],
      "responses": {}
    }
  },
  "/api/mobile/recipe/comment/list": {
    "get": {
      "x-swagger-router-controller": "/api/mobile/recipe/comment/list",
      "operationId": "index",
      "tags": ["All Comments listing "],
      "description": "",
      "parameters": [
        { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
        { "name": "recipeId",  "in": "query", "type":"string","required":true},
        { "name": "page",  "in": "query", "type":"string"},
        { "name": "limit",  "in": "query", "type":"string"}

       ],
      "responses": {}
    }
  },
...
  }
}